fix + update
- hoàn thành chỉnh sửa câu e
- chưa lưu bill vào file
cập nhật :
+  hiển thị số lượng tồn khi số lượng xuất > số lượng tồn
+ chỉnh sửa số lượng tồn có trong cây khi nhập xong bill
+fix lỗi crash chương trình

+++ Project/include/Functions.cpp
@@ -20,9 +20,9 @@ bool CheckNum(char s)
 	return 0;	 
 }
 
-int ChangeCharToNum (char s[])
+long ChangeCharToNum (char s[])
 {
-	int n = 0;
+	long n = 0;
 	for(int i =0; i<strlen(s);i++)
 	{
 		n *= 10;
